list of pydantic types
